Add deployment configuration using fly.io
The frontend is hosted here: https://tele-tale.fly.dev/
The backend is hotsed here: https://tele-tale-api.fly.dev/

For now, the deployment must be done manually by running the
`fly deploy`, command. We only need to do so when we want to
evenually make a demo video of people playing the game. Right
now, when deployed to localhost, we can't have actual multiplayer.